# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Anatomiya",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/lesson-join-object-file": {
            "get": {
                "tags": [
                    "Lesson_Join_Object_File"
                ],
                "summary": "All lesson join object File",
                "description": "Hamma lesson join object filelarlarni ko'rish",
                "operationId": "lesson_join_object_file_index",
                "responses": {
                    "200": {
                        "description": "Successful",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": "1"
                                        },
                                        "lesson_id": {
                                            "type": "integer",
                                            "example": "5"
                                        },
                                        "object_file_id": {
                                            "type": "integer",
                                            "example": "4"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Lesson_Join_Object_File"
                ],
                "summary": "lesson_join_object File add",
                "description": "lesson join object filelarlar qo'shish",
                "operationId": "lesson_join_object_file_store",
                "requestBody": {
                    "description": "lesson join object file save",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "lesson_id",
                                    "object_file_id"
                                ],
                                "properties": {
                                    "lesson_id": {
                                        "type": "integer",
                                        "format": "number",
                                        "example": "6"
                                    },
                                    "object_file_id": {
                                        "type": "integer",
                                        "format": "number",
                                        "example": "2"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": "1"
                                        },
                                        "lesson_id": {
                                            "type": "integer",
                                            "example": "Harley"
                                        },
                                        "object_file_id": {
                                            "type": "integer",
                                            "example": "Harley"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/lesson-join-object-file/{id}": {
            "get": {
                "tags": [
                    "Lesson_Join_Object_File"
                ],
                "summary": "Show lesson join object file",
                "description": "bitta lesson join object fileda hamma malumotlarini ko'rsatadi",
                "operationId": "lesson_Join_Object_File_show",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "lesson_join_object file id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": "1"
                                        },
                                        "lesson_id": {
                                            "type": "integer",
                                            "example": "Harley"
                                        },
                                        "object_file_id": {
                                            "type": "integer",
                                            "example": "Harley"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Lesson_Join_Object_File"
                ],
                "summary": "Update existing project",
                "description": "Returns updated project data",
                "operationId": "lesson_join_object_file_update",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "lesson_join_object file id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "description": "term update",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "lesson_id",
                                    "object_file_id"
                                ],
                                "properties": {
                                    "lesson_id": {
                                        "type": "integer",
                                        "format": "number",
                                        "example": "6"
                                    },
                                    "object_file_id": {
                                        "type": "integer",
                                        "format": "number",
                                        "example": "2"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": "1"
                                        },
                                        "lesson_id": {
                                            "type": "integer",
                                            "example": "Harley"
                                        },
                                        "object_file_id": {
                                            "type": "integer",
                                            "example": "Harley"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Lesson_Join_Object_File"
                ],
                "summary": "Delete existing project",
                "description": "Deletes a record and returns no content",
                "operationId": "lesson_join_object_file_delete",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Project id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    },
...
    "security": {
        "bearer_token": {
            "type": "apiKey",
            "description": "Enter token in format (Bearer <token>)",
            "name": "Authorization",
            "in": "header"
        }
    }
}